% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR2.R
\name{getR2}
\alias{getR2}
\title{getR2}
\usage{
getR2(x, z, model)
}
\arguments{
\item{x}{Gene expression matrix, n*p (rows for subjects and columns for genes).}

\item{z}{One phenotypic variable from clinical dataset, a vector.}

\item{model}{The model fitted to obtain R2, please select model from 'linear', 'logit', 'exp', 'polr','cox'.}
}
\value{
A vector
\item{R2.per}{R-squared or pseudo R-squared between phenotypic variable and expression value of each gene, a vector.}
}
\description{
Get R-squared or pseudo R-squared
}
\details{
Obtain R-squared or pseudo R-squared between phenotypic variable and expression value of each gene,
which will be used in GuidedSparseKmeans.R2out function.
}
\author{
Lingsong Meng
}
