% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuidedSparseKmeans.KLam.R
\name{GuidedSparseKmeans.KLam}
\alias{GuidedSparseKmeans.KLam}
\title{GuidedSparseKmeans.KLam}
\usage{
GuidedSparseKmeans.KLam(x, z, pre.K = NULL, s.one, model, nstart = 20,
  maxiter = 15, silence = F)
}
\arguments{
\item{x}{Gene expression matrix, n*p (rows for subjects and columns for genes).}

\item{z}{One phenotypic variable from clinical dataset, a vector.}

\item{pre.K}{Pre-knowledge of the number of clusters.}

\item{s.one}{A proper value of the boundary of l1n weights.}

\item{model}{The model fitted to obtain R2, please select model from 'linear', 'logit', 'exp', 'polr','cox'.}

\item{nstart}{Specify the number of starting point for K-means.}

\item{maxiter}{Maximum number of iteration.}

\item{silence}{Output progress or not.}
}
\value{
A list consisting of
\item{K.select}{value of selected K.}
\item{lam.select}{value of selected lam.}
\item{R2.per}{R-squared or pseudo R-squared between phenotypic variable and expression value of each gene, a vector.}
\item{ARI.Cs}{Adjusted ARI values for cluster results.}
\item{Jaccard.gene}{Jaccard index values for gene selection results.}
}
\description{
Selection of Tuning Parameter K and lam in Guided Sparse K-means
}
\details{
Select tuning parameter K using gap statistics and tuning parameter lam using sensitivity analysis in Guided Sparse K-means integrating clinical dataset with gene expression dataset.
}
\author{
Lingsong Meng
}
