% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuidedSparseKmeans.R
\name{GuidedSparseKmeans}
\alias{GuidedSparseKmeans}
\title{GuidedSparseKmeans}
\usage{
GuidedSparseKmeans(x, z, K, s, lam, model, nstart = 20, maxiter = 15,
  silence = F)
}
\arguments{
\item{x}{Gene expression matrix, n*p (rows for subjects and columns for genes).}

\item{z}{One phenotypic variable from clinical dataset, a vector.}

\item{K}{Number of clusters.}

\item{s}{The boundary of l1n weights, a vector.}

\item{lam}{The intensity of guidance.}

\item{model}{The model fitted to obtain R2, please select model from 'linear', 'logit', 'exp', 'polr','cox'.}

\item{nstart}{Specify the number of starting point for K-means.}

\item{maxiter}{Maximum number of iteration.}

\item{silence}{Output progress or not.}
}
\value{
m lists, m is the length of parameter s. Each list is consisting of
\item{weights}{weight for each feature, zero weight means the feature is not selected.}
\item{clusters}{cluster results.}
\item{object}{objective value.}
\item{bound}{a boundary of l1n weights}
\item{R2.per}{R-squared or pseudo R-squared between phenotypic variable and expression value of each gene, a vector.}
}
\description{
Guided Sparse K-means
}
\details{
Guided Sparse K-means integrating clinical dataset with gene expression dataset.
}
\author{
Lingsong Meng
}
