% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuidedSparseKmeans.S.R2out.R
\name{GuidedSparseKmeans.S.R2out}
\alias{GuidedSparseKmeans.S.R2out}
\title{GuidedSparseKmeans.S}
\usage{
GuidedSparseKmeans.S.R2out(x, R2.per, K, s, lam, nstart = 20,
  maxiter = 15, nperms = 50, silence = F)
}
\arguments{
\item{x}{Gene expression matrix, n*p (rows for subjects and columns for genes).}

\item{R2.per}{R-squared or pseudo R-squared between phenotypic variable and expression value of each gene, a vector.}

\item{K}{Number of clusters.}

\item{s}{The boundary of l1n weights, a vector.}

\item{lam}{The intensity of guidance.}

\item{nstart}{Specify the number of starting point for K-means.}

\item{maxiter}{Maximum number of iteration.}

\item{nperms}{Number of permutation times}

\item{silence}{Output progress or not.}
}
\value{
A list consisting of
\item{nnonzerows}{number of nonzero weights.}
\item{gaps}{gap statistics.}
\item{segaps}{statard error of gap statistics.}
\item{s}{candidates of s.}
\item{s.best}{the best s with largest gap.}
}
\description{
Selection of Tuning Parameter s in Guided Sparse K-means
}
\details{
Select tuning parameter s via permutation in Guided Sparse K-means integrating clinical dataset with gene expression dataset.
}
\author{
Lingsong Meng
}
