% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuidedSparseKmeans.S.R
\name{GuidedSparseKmeans.S}
\alias{GuidedSparseKmeans.S}
\title{GuidedSparseKmeans.S}
\usage{
GuidedSparseKmeans.S(x, z, K, s, lam, model, nstart = 20, maxiter = 15,
  nperms = 50, silence = F)
}
\arguments{
\item{x}{Gene expression matrix, n*p (rows for subjects and columns for genes).}

\item{z}{One phenotypic variable from clinical dataset, a vector.}

\item{K}{Number of clusters.}

\item{s}{The boundary of l1n weights, a vector.}

\item{lam}{The intensity of guidance.}

\item{model}{The model fitted to obtain R2, please select model from 'linear', 'logit', 'exp', 'polr','cox'.}

\item{nstart}{Specify the number of starting point for K-means.}

\item{maxiter}{Maximum number of iteration.}

\item{nperms}{Number of permutation times}

\item{silence}{Output progress or not.}
}
\value{
A list consisting of
\item{nnonzerows}{number of nonzero weights.}
\item{gaps}{gap statistics.}
\item{segaps}{statard error of gap statistics.}
\item{s}{candidates of s.}
\item{s.best}{the best s with largest gap.}
}
\description{
Selection of Tuning Parameter s in Guided Sparse K-means
}
\details{
Select tuning parameter s via permutation in Guided Sparse K-means integrating clinical dataset with gene expression dataset.
}
\author{
Lingsong Meng
}
